# backend/docs/swagger.yaml
basePath: /
definitions:
...
  router.swaggerErrorResponse:
    properties:
      error:
        description: 错误信息
        example: 请求参数错误
        type: string
    type: object
  router.swaggerGenericObject:
    additionalProperties: true
    type: object
...
  router.swaggerIDResponse:
    properties:
      id:
        description: 编号
        example: 1
        type: integer
    type: object
  router.swaggerImportDatabaseResponse:
    properties:
      message:
        description: 消息
        example: 数据库导入成功
        type: string
      ok:
        description: 是否成功
        example: true
        type: boolean
    type: object
...
  router.swaggerSystemBaseConfig:
    properties:
      appName:
        description: 控制台品牌名称
        example: ITDB
        type: string
      appSubtitle:
        description: 控制台品牌副标题
        example: IT Asset Management
        type: string
      backupCron:
        description: 定时备份 Cron 表达式
        example: 0 0 * * *
        type: string
      backupEnabled:
        description: 是否启用定时备份
        type: boolean
      backupRetentionDays:
        description: 备份保留天数（0 为永久）
        example: 30
        type: integer
      iconData:
        description: 图标（站点路径或 Data URL）
        example: /favicon.svg
        type: string
      loginName:
        description: 登录页品牌名称
        example: ITDB
        type: string
      passwordResetRateLimitMinutes:
        description: 发送限流统计窗口（分钟）
        example: 5
        type: integer
      passwordResetSendCooldownMinutes:
        description: 验证码发送冷却（分钟）
        example: 0.5
        type: number
      resetCaptchaTtlMinutes:
        description: 图形验证码有效期（分钟）
        example: 1
        type: integer
      resetCodeTtlMinutes:
        description: 找回密码验证码有效期（分钟）
        example: 10
        type: integer
      section:
        example: brand
        type: string
      siteName:
        description: 网站名称
        example: ITDB
        type: string
    type: object
...
info:
  contact: {}
  description: ITDB IT 资产管理系统后端 API 文档。除登录和健康检查外，接口默认需要 Bearer Token。
  title: ITDB API
  version: 1.0.0
paths:
...
  /api/backups/database:
    get:
      description: 下载数据库备份（itdb-日期.db）；携带 files=1 时按数据库实际引用的文件打包为 zip（文件位于包内 files
        目录），引用文件全部不存在或未引用时仅导出 db 文件，缺失文件写入备份审计
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: 下载数据库备份
      tags:
      - 备份
...
  /api/files:
    get:
      description: 获取文件列表，支持关键字搜索
      parameters:
      - description: 搜索关键词
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/router.swaggerGenericObject'
            type: array
      security:
      - BearerAuth: []
      summary: 获取文件列表
      tags:
      - 文件
    post:
      consumes:
      - multipart/form-data
      description: 上传文件（multipart），支持关联到硬件、软件、单据、合同
      parameters:
      - description: 标题
        in: formData
        name: title
        required: true
        type: string
      - description: 文件类型 ID
        in: formData
        name: typeId
        required: true
        type: integer
      - description: 签署日期
        in: formData
        name: date
        type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: 关联硬件 ID，逗号分隔
        in: formData
        name: itemLinks
        type: string
      - description: 关联软件 ID，逗号分隔
        in: formData
        name: softwareLinks
        type: string
      - description: 关联发票 ID，逗号分隔
        in: formData
        name: invoiceLinks
        type: string
      - description: 关联合同 ID，逗号分隔
        in: formData
        name: contractLinks
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/router.swaggerIDResponse'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - 文件
...
  /api/import/database:
    post:
      consumes:
      - multipart/form-data
      description: 上传 .db 或 zip 替换当前数据库，压缩包内的上传文件恢复到上传目录，导入前自动备份；旧版平台数据库自动转换，仅迁移资产管理、资料管理与用户数据，系统配置、审计历史、标签预设保持当前默认
      parameters:
      - description: SQLite .db 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.swaggerImportDatabaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: 导入 SQLite 数据库
      tags:
      - 数据库导入
...
  /api/settings/base:
    get:
      description: 获取系统基础配置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.swaggerSystemBaseConfig'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/router.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取系统基础配置
      tags:
      - 系统设置
    put:
      consumes:
      - application/json
      description: 更新系统基础配置；携带 section（brand/security/backup）时按区块局部更新，审计目标细分为品牌标识、安全时效、数据备份
      parameters:
      - description: 系统基础配置
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.swaggerSystemBaseConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.swaggerSystemBaseConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.swaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/router.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新系统基础配置
      tags:
      - 系统设置
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"